Create MovieManager_nosys.c
This file is the have the same functions as MovieManager.c file but instead of system("clear"); Which is incompatible with windows devices, This file has printf("\n\n"); to help with readability. except the change that stated above this file does the same things as MovieManager.c and has the same restrictions. This not a update on MovieManager program.